Change the constants corresponding to DNSSEC algorithms and security status from byte to int.
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1161 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/Verifier.java
@@ -19,6 +19,6 @@ public interface Verifier {
  * @see RRset
  * @see Cache
  */
-byte verify(RRset set, Cache cache);
+int verify(RRset set, Cache cache);
 
 }
